# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend vor Rupay",
        "title": "Rupay Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api",
    "paths": {
...
        "/transaction": {
            "get": {
                "description": "Retrieve a list of all transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Retrieve all transactions",
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Transaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new transaction with the provided date and price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "CreateTransaction payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction data",
                        "schema": {
                            "$ref": "#/definitions/db.Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a transaction with the provided date and price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Update a transaction",
                "parameters": [
                    {
                        "description": "UpdateTransaction payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction data",
                        "schema": {
                            "$ref": "#/definitions/db.Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transaction/{id}": {
            "get": {
                "description": "Retrieve a transaction by the provided id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Retrieve a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction data",
                        "schema": {
                            "$ref": "#/definitions/db.Transaction"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a transaction by the provided id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaction deleted successfully"
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete Transaction",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "db.Transaction": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "e.ErrorDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field causing the error",
                    "type": "string"
                },
                "issue": {
                    "description": "Description of the issue with the field",
                    "type": "string"
                }
            }
        },
        "e.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string"
                },
                "details": {
                    "description": "List of field-specific errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/e.ErrorDetail"
                    }
                },
                "error": {
                    "description": "Stack Trace",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                }
            }
        },
...
        "schemas.CreateTransaction": {
            "type": "object",
            "required": [
                "date",
                "price"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2024-01-24T00:00:00Z"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "schemas.UpdateTransaction": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2024-01-24T00:00:00Z"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
    }
}